requirements.txt
aiohttp==3.7.4.post0
aioredis==2.0.1
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
chardet==4.0.0
discord==2.1.0
discord.py==2.1.0
emojis==0.6.0
idna==3.3
multidict==6.0.2
Pillow==9.3.0
pycodestyle==2.8.0
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
toml==0.10.2
typing_extensions==4.0.1
yarl==1.7.2
